remove javax dep, add licences, move strings to constants
