+++ addons/monitoring-addon/monitoring-core/monitoring-core-quarkus/src/main/resources/META-INF/kogito.addon
@@ -1,0 +1,1 @@
+monitoring
