+++ addons/monitoring-addon/monitoring-core/monitoring-core-springboot/pom.xml
@@ -24,10 +24,6 @@
 
   <dependencies>
     <dependency>
-      <groupId>javax.annotation</groupId>
-      <artifactId>javax.annotation-api</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.kie.kogito</groupId>
       <artifactId>monitoring-core-common</artifactId>
     </dependency>
