+++ addons/monitoring-addon/monitoring-core/monitoring-core-springboot/src/main/resources/META-INF/kogito.addon
@@ -1,0 +1,1 @@
+monitoring
