+++ addons/monitoring-addon/monitoring-core/pom.xml~HEAD
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <artifactId>monitoring-addon</artifactId>
+    <groupId>org.kie.kogito</groupId>
+    <version>2.0.0-SNAPSHOT</version>
+  </parent>
+  <name>Kogito :: Add-Ons :: Monitoring Core</name>
+
+  <description>Core Module for monitoring addon</description>
+  <artifactId>monitoring-core</artifactId>
+
+  <packaging>pom</packaging>
+
+  <modules>
+    <module>monitoring-core-common</module>
+    <module>monitoring-core-quarkus</module>
+    <module>monitoring-core-springboot</module>
+  </modules>
+
+</project>
