+++ addons/monitoring-addon/monitoring-prometheus-addon/monitoring-prometheus-common/pom.xml
@@ -24,12 +24,6 @@
       <artifactId>micrometer-registry-prometheus</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>org.jboss.spec.javax.ws.rs</groupId>
-      <artifactId>jboss-jaxrs-api_2.1_spec</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
     <!-- test dependencies -->
     <dependency>
       <groupId>org.junit.jupiter</groupId>
