+++ addons/monitoring-addon/monitoring-prometheus-addon/monitoring-prometheus-common/src/test/resources/META-INF/kogito.addon
@@ -1,0 +1,1 @@
+prometheus-monitoring
